make sure GOROOT is set (fixes #13)
because the invoked gcov2lcov uses golang API functions that rely on
GOROOT (i.e. build.Import, which calls
https://cs.opensource.google/go/go/+/refs/tags/go1.18.2:src/go/build/build.go;drc=0a1a092c4b56a1d4033372fbd07924dad8cbb50b;l=544)

Up to version v3 of the github setup-go action, the GOROOT variable was
automatically set. But that changed with v3 of the setup-go action, and
that's why we now set GOROOT.